Add normalized_walk function to DeepScan
Prior to this commit normalizing was done in a function that was called
in a busy loop. That function call checked `sys.platform` and then
normalized a filename. This commit moves the function call and
`sys.platform` check out of the loop by creating a new `normalized_walk`
function which is like `os.walk` but does the normalization efficiently.